requirements.txt
langchain-openai>=0.1.0
langgraph>=0.0.40
PyPDF2>=3.0.1
python-dotenv>=1.0.0
streamlit>=1.28.0
sqlalchemy>=2.0.0
pandas>=2.0.0
tiktoken>=0.5.0
